feat: Added spotless copyright templates; renamed old files
