+++ apps/wearos/src/androidTest/kotlin/dev/marlonlom/mocca/wearos/features/calculator/input/CalculatorInputScreenUiTest.kt
@@ -3,12 +3,13 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package dev.marlonlom.mocca.wearos.features.calculator_input
+package dev.marlonlom.mocca.wearos.features.input
 
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.performClick
 import com.google.common.truth.Truth.assertThat
+import dev.marlonlom.mocca.wearos.features.calculator.input.CalculatorInputScreen
 import org.junit.Rule
 import org.junit.Test
