+++ apps/wearos/src/androidTest/kotlin/dev/marlonlom/mocca/wearos/features/calculator/output/FailureCalculatorOutputUiTest.kt
@@ -3,12 +3,13 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package dev.marlonlom.mocca.wearos.features.calculator_output
+package dev.marlonlom.mocca.wearos.features.output
 
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.ui.test.assertIsDisplayed
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithText
+import dev.marlonlom.mocca.wearos.features.calculator.output.FailureCalculatorOutput
 import org.junit.Rule
 import org.junit.Test
