+++ apps/wearos/src/androidTest/kotlin/dev/marlonlom/mocca/wearos/features/calculator/output/SuccessCalculatorOutputUiTest.kt
@@ -3,13 +3,14 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package dev.marlonlom.mocca.wearos.features.calculator_output
+package dev.marlonlom.mocca.wearos.features.output
 
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.ui.test.assertIsDisplayed
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithText
 import dev.marlonlom.mocca.calculator.model.CalculationResult
+import dev.marlonlom.mocca.wearos.features.calculator.output.SuccessCalculatorOutput
 import org.junit.Rule
 import org.junit.Test
