+++ apps/wearos/src/main/kotlin/dev/marlonlom/mocca/wearos/features/calculator/output/FailureCalculatorOutput.kt
