+++ spotless/copyright.kt
@@ -1,0 +1,4 @@
+/*
+ * Copyright 2024 Marlonlom
+ * SPDX-License-Identifier: Apache-2.0
+ */
